+++ js/karma-coverage.js
@@ -1,8 +1,8 @@
 const percentage = {
-    lines: 96,
-    statements: 96,
-    functions: 83,
-    branches: 96
+    lines: 100,
+    statements: 100,
+    functions: 100,
+    branches: 100
 }
 var summary = require('./karma/coverage/coverage-summary.json');