+++ src/yafowil/widget/ace/resources/widget.js
@@ -5,6 +5,10 @@ var yafowil_ace = (function (exports, $) {
         static initialize(context) {
             $('.ace-editor-wrapper', context).each(function() {
                 let elem = $(this);
+                if (window.yafowil_array !== undefined &&
+                    window.yafowil_array.inside_template(elem)) {
+                    return;
+                }
                 new AceWidget(elem, elem.data('yafowil-ace'));
             });
         }
@@ -27,6 +31,15 @@ var yafowil_ace = (function (exports, $) {
             this.textarea.val(this.editor.getValue());
         }
     }
+    function ace_on_array_add(inst, context) {
+        AceWidget.initialize(context);
+    }
+    function register_array_subscribers() {
+        if (window.yafowil_array === undefined) {
+            return;
+        }
+        window.yafowil_array.on_array_event('on_add', ace_on_array_add);
+    }
 
     $(function() {
         if (window.ts !== undefined) {
@@ -36,9 +49,12 @@ var yafowil_ace = (function (exports, $) {
         } else {
             AceWidget.initialize();
         }
+        register_array_subscribers();
     });
 
     exports.AceWidget = AceWidget;
+    exports.ace_on_array_add = ace_on_array_add;
+    exports.register_array_subscribers = register_array_subscribers;
 
     Object.defineProperty(exports, '__esModule', { value: true });