+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 2.0 (unreleased)
 ----------------
 
+- Extend JS by ``ace_on_array_add`` and ``register_array_subscribers``
+  to enable usage in ``yafowil.widget.array``.
+  [lenadax]
+
 - Update ACE Editor to v1.4.13.
   [rnix]